bug: add timeout version check in rewards and make fn reusable
